makefiles: refactoring makefile
* Drop MCUFIRMWARE_PLATFORM_BASE: a platform is a module
* platforms/cortex module name is pf_cortex
* Drop Makefile.dep as it is handled by RIOT
* Include $(APPDIR)/include/* only if it exists
* Use RIOT EXTERNAL_MODULE_DIRS philosophy
