--- Makefile.dep
@@ -1,12 +1,0 @@
-# Dependencies makefiles are included in a specific order
-
-# First platfoms as they depend on other modules, hardware or software
-include $(MCUFIRMWAREBASE)/platforms/Makefile.dep
-
-# Next pure software components that could depend on hardware drivers
-include $(MCUFIRMWAREBASE)/controllers/Makefile.dep
-include $(MCUFIRMWAREBASE)/planners/Makefile.dep
-include $(MCUFIRMWAREBASE)/sys/Makefile.dep
-
-# Finally hardware drivers that could depend on each other
-include $(MCUFIRMWAREBASE)/drivers/Makefile.dep
