+++ applications/cogip2020-cortex/Makefile
@@ -2,7 +2,7 @@ APPLICATION = cortex
 
 BOARD ?= cogip2019-cortex
 
-MCUFIRMWARE_PLATFORM_BASE = cortex
+USEMODULE += pf_cortex
 
 # Main thread stack size
 CFLAGS += -DTHREAD_STACKSIZE_MAIN=4096
