+++ drivers/Makefile
@@ -1,3 +1,1 @@
-DIRS += $(dir $(wildcard $(addsuffix /Makefile, $(USEMODULE))))
-
 include $(RIOTBASE)/Makefile.base
