+++ drivers/Makefile.include
@@ -1,2 +1,6 @@
-DIRS += $(MCUFIRMWAREBASE)/drivers
-INCLUDES += -I$(MCUFIRMWAREBASE)/drivers/include/
+LAST_MAKEFILEDIR_drivers := $(LAST_MAKEFILEDIR)
+
+# Drivers
+EXTERNAL_MODULE_DIRS += $(LAST_MAKEFILEDIR_drivers)/lds01
+EXTERNAL_MODULE_DIRS += $(LAST_MAKEFILEDIR_drivers)/pca9548
+EXTERNAL_MODULE_DIRS += $(LAST_MAKEFILEDIR_drivers)/sd21
