+++ drivers/pca9548/Makefile
@@ -1,7 +1,3 @@
 MODULE = pca9548
 
-ifneq (,$(filter calibration,$(MCUFIRMWARE_OPTIONS)))
-	DIRS += calibration
-endif
-
 include $(RIOTBASE)/Makefile.base
