+++ drivers/pca9548/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_pca9548 := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_pca9548)
