+++ drivers/sd21/Makefile
@@ -1,7 +1,3 @@
 MODULE = sd21
 
-ifneq (,$(filter calibration,$(MCUFIRMWARE_OPTIONS)))
-	DIRS += calibration
-endif
-
 include $(RIOTBASE)/Makefile.base
