+++ drivers/sd21/Makefile.include
@@ -1,0 +1,3 @@
+USEMODULE_INCLUDES_sd21 := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_sd21)
+
