+++ planners/Makefile
@@ -1,7 +1,1 @@
-MODULE = planner
-
-ifneq (,$(filter calibration,$(MCUFIRMWARE_OPTIONS)))
-	DIRS += calibration
-endif
-                                                                                
 include $(RIOTBASE)/Makefile.base
