+++ planners/astar/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_astar := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_astar)
