+++ planners/shell_pl/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE_INCLUDES_shell_pl := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_shell_pl)
