+++ planners/shell_pl/calib_planner.c
@@ -10,8 +10,8 @@
 /* Project includes */
 #include "planner.h"
 #include "platform.h"
-#include "calibration/calib_planner.h"
-#include "calibration/calib_platform.h"
+#include "calib_planner.h"
+#include "calib_platform.h"
 
 /* Shell command array */
 static shell_command_linked_t pln_shell_commands;
