+++ platforms/Makefile
@@ -1,7 +1,1 @@
-DIRS += $(MCUFIRMWARE_PLATFORM_BASE)
-
-ifneq (,$(filter calibration,$(MCUFIRMWARE_OPTIONS)))
-	DIRS += calibration
-endif
-
 include $(RIOTBASE)/Makefile.base
