--- platforms/Makefile.dep
@@ -1,3 +1,0 @@
-USEMODULE += $(MCUFIRMWARE_PLATFORM_BASE)
-
-include $(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)/Makefile.dep
