+++ platforms/Makefile.include
@@ -1,9 +1,7 @@
-DIRS += $(MCUFIRMWAREBASE)/platforms
-DIRS += $(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)
+LAST_MAKEFILEDIR_platforms := $(LAST_MAKEFILEDIR)
 
-INCLUDES += -I$(MCUFIRMWAREBASE)/platforms/include/
-INCLUDES += -I$(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)/include
+# Calibration
+EXTERNAL_MODULE_DIRS += $(LAST_MAKEFILEDIR_platforms)/shell_pf
 
-CFLAGS += -DMCUFIRMWARE_PLATFORM_BASE=\"$(MCUFIRMWARE_PLATFORM_BASE)\"
-
-include $(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)/Makefile.include
+# Platforms
+EXTERNAL_MODULE_DIRS += $(LAST_MAKEFILEDIR_platforms)/cortex
