+++ platforms/cortex/Makefile.dep
@@ -11,7 +11,7 @@ USEMODULE += xtimer
 # mcu-firmware modules
 USEMODULE += ctrl
 USEMODULE += pca9548
-USEMODULE += planner
+USEMODULE += pl_astar
 USEMODULE += quadpid
 USEMODULE += robotics
 USEMODULE += sd21
@@ -26,6 +26,6 @@ FEATURES_REQUIRED += periph_qdec
 FEATURES_REQUIRED += periph_pwm
 
 ifneq (,$(filter calibration,$(MCUFIRMWARE_OPTIONS)))
-	USEMODULE += calib_platform
+	USEMODULE += shell_pf
 	USEMODULE += fmt
 endif
