+++ platforms/cortex/Makefile.include
@@ -1,1 +1,2 @@
-DEVELHELP=1
+USEMODULE_INCLUDES_cortex := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_cortex)
