+++ platforms/shell_pf/Makefile.dep
@@ -1,0 +1,1 @@
+USEMODULE += shell_pl
