+++ platforms/shell_pf/Makefile.include
@@ -1,0 +1,4 @@
+USEMODULE_INCLUDES_calibration := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_calibration)
+
+
